Refactor timer: remove useless sync with mutex+cond_var, start is now a wrapper on init for retrocompatibility
