+++ lib/include/kickcat/OS/Timer.h
@@ -1,12 +1,9 @@
 #ifndef KICKCAT_OS_TIMER_H
 #define KICKCAT_OS_TIMER_H
 
-#include <string_view>
 #include <system_error>
 
 #include "kickcat/OS/Time.h"
-#include "kickcat/OS/ConditionVariable.h"
-#include "kickcat/OS/Mutex.h"
 
 namespace kickcat
 {
@@ -17,19 +14,14 @@ namespace kickcat
         Timer(Timer const &) = delete;
         void operator=(Timer const &) = delete;
 
-        /// \param  name     Name of the timer (logging)
         /// \param  period   Interval between two timer firing
         Timer(nanoseconds period);
         ~Timer() = default;
 
-        /// Start the timer.
+        /// \brief  Start (or restart) the timer.
+        /// \details Aligns the next deadline to sync_point + N*period so that it lies in the future.
+        ///          Safe to call multiple times.
         void start(nanoseconds sync_point = since_epoch());
-
-        ///  Stop the timer.
-        void stop();
-
-        ///  Get timer status
-        bool is_stopped() const;
 
         ///  Change the timer values to the new provided values and reset it.
         ///  This does not take effect before the timer is restarted.
@@ -42,16 +34,10 @@ namespace kickcat
         /// Wait until the next timer tick (blocking call)
         std::error_code wait_next_tick();
 
-    protected:
     private:
-        std::string name_;
         nanoseconds period_;
         nanoseconds next_deadline_{};
         nanoseconds last_wakeup_{};
-
-        Mutex mutex_{};
-        ConditionVariable stop_{};
-        bool is_stopped_{true};
     };
 }
