+++ lib/src/OS/Unix/Timer.cc
@@ -1,4 +1,3 @@
-#include <cstdio>
 #include <stdexcept>
 #include <unistd.h>
 
@@ -8,63 +7,8 @@
 
 namespace kickcat
 {
-    Timer::Timer(nanoseconds period)
-        : period_{period}
-    {
-    }
-
-    nanoseconds Timer::period() const
-    {
-        return period_;
-    }
-
-    void Timer::start(nanoseconds sync_point)
-    {
-        nanoseconds now = since_epoch();
-        nanoseconds delta = now - sync_point;
-
-        int64_t periods_to_skip = 0;
-        if (delta >= 0ns)
-        {
-            periods_to_skip = (delta / period_) + 1;
-        }
-        next_deadline_ = sync_point + periods_to_skip * period_;
-
-        {
-            LockGuard lock(mutex_);
-            is_stopped_ = false;
-        }
-        stop_.signal();
-    }
-
-    void Timer::stop()
-    {
-        {
-            LockGuard lock(mutex_);
-            is_stopped_ = true;
-        }
-        stop_.signal();
-    }
-
-    bool Timer::is_stopped() const
-    {
-        return is_stopped_;
-    }
-
-    void Timer::update_period(nanoseconds period)
-    {
-        period_ = period;
-    }
-
     std::error_code Timer::wait_next_tick()
     {
-        {
-            LockGuard lock(mutex_);
-            stop_.wait(mutex_, [&]()
-                       { return not is_stopped_; });
-        }
-
-        // Wait to the next working time.
         timespec const deadline = to_timespec(next_deadline_);
         int rc = clock_nanosleep(CLOCK_REALTIME, TIMER_ABSTIME, &deadline, NULL);
         if (rc != 0)
@@ -94,7 +38,6 @@ namespace kickcat
             // We are late: compute a new deadline that maintain the cycle.
             next_deadline_ += period_;
         }
-
 
         return ret;
     }
